backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI/ML
#tensorflow-cpu==2.13.0
#keras==2.15.0
numpy>=1.26.4
#pandas==2.2.0
#scikit-learn==1.4.0
#shap==0.44.1

# Firebase
firebase-admin==6.4.0

# Google APIs
google-generativeai==0.3.2
googlemaps==4.10.0

# Utilities
python-dotenv==1.0.0
- pydantic==2.5.3
+ pydantic==1.10.13
cryptography==42.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
httpx==0.26.0
